mildly rearrange guts of image.rs to remove all intra-crate dependencies, preparing it for excision
