+++ crates/bevy_render/src/mesh/primitives/dim3/plane.rs
@@ -3,7 +3,7 @@ use wgpu::PrimitiveTopology;
 
 use crate::{
     mesh::{Indices, Mesh, MeshBuilder, Meshable},
-    render_asset::RenderAssetUsages,
+    texture::RenderAssetUsages,
 };
 
 /// A builder used for creating a [`Mesh`] with a [`Plane3d`] shape.
