+++ crates/bevy_render/src/texture/hdr_texture_loader.rs
@@ -1,5 +1,5 @@
 use crate::{
-    render_asset::RenderAssetUsages,
+    texture::image::RenderAssetUsages,
     texture::{Image, TextureFormatPixelInfo},
 };
 use bevy_asset::{io::Reader, AssetLoader, LoadContext};
