+++ crates/bevy_render/src/texture/image_texture_conversion.rs
@@ -1,5 +1,5 @@
 use crate::{
-    render_asset::RenderAssetUsages,
+    texture::image::RenderAssetUsages,
     texture::{Image, TextureFormatPixelInfo},
 };
 use image::{DynamicImage, ImageBuffer};
